discord_agent/requirements.txt
# Основные зависимости
discord.py>=2.3.2
requests>=2.31.0
python-dotenv>=1.0.0

# Для разработки (опционально)
# ipython>=8.0.0
# black>=23.0.0
